feat: service default is working properly
