+++ src/Customers/Customers/Customers.csproj
@@ -15,7 +15,7 @@
   </ItemGroup>
 
     <ItemGroup>
-    <ProjectReference Include="..\..\Trinity.ServiceDefault\Trinity.ServiceDefault.csproj" />
+    <ProjectReference Include="..\..\Aspire\Trinity.ServiceDefaults\Trinity.ServiceDefaults.csproj" />
   </ItemGroup>
 
     <ItemGroup>
