+++ src/Customers/Customers/Program.cs
@@ -5,16 +5,9 @@ var builder = WebApplication.CreateBuilder(args);
 
 builder.AddServiceDefaults();
 
- builder.ConfigureSerilog();
- builder.ConfigureScalar();
- builder.AddDefaultHealthChecks();
-
 var app = builder.Build();
 
-
- app.AddScalar();
-
- app.MapDefaultEndpoints();
+app.MapDefaultEndpoints();
 
 app.UseHttpsRedirection();
