--- src/Trinity.ServiceDefault/Trinity.ServiceDefault.sln
@@ -1,25 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.5.002.0
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Trinity.ServiceDefault", "Trinity.ServiceDefault.csproj", "{C828D871-F9BF-4BCF-BC62-CB24118D876C}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{C828D871-F9BF-4BCF-BC62-CB24118D876C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{C828D871-F9BF-4BCF-BC62-CB24118D876C}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{C828D871-F9BF-4BCF-BC62-CB24118D876C}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{C828D871-F9BF-4BCF-BC62-CB24118D876C}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {3DDDA0F3-A47F-42AE-B3E4-E33F40A2106D}
-	EndGlobalSection
-EndGlobal
